requirements.txt
confluent-kafka==2.1.1
apache-airflow-providers-apache-kafka==1.0.0
apache-airflow-providers-postgres>=5.0.0
psycopg2-binary>=2.9.0
faker==19.13.0
kafka-python==2.0.2